pyproject.toml
[tool.poetry]
name = "rplay-live-dl"
version = "1.0.0"
description = ""
authors = ["paver(kevinsun)"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
pydantic = "^2.10.4"
yt-dlp = "^2024.12.23"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
apscheduler = "^3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
